Medifast/requirements.txt
certifi==2018.8.24
dj-database-url==0.5.0
django==2.1.1
django-heroku==0.3.1
gunicorn==19.9.0
pipenv==2018.7.1
psycopg2==2.7.5
pytz==2018.5
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==4.1
djangorestframework==3.8.2
